Refactor and do cleanup
